pyproject.toml
[project]
name = "pytestbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "discord-py>=2.5.2",
    "pynacl>=1.5.0",
    "requests>=2.32.3",
    "yt-dlp>=2025.3.31",
]
